requirements.txt
Sphinx==1.7.6
sphinx_rtd_theme==0.4.0
pylint==2.0.1
pytest==3.6.3
pytest-cov==2.5.1
cachetools
parameterized==0.6.1
pyyaml